+++ src/overlay100/ov100_021D13E4.c
@@ -46,7 +46,7 @@ void *ov100_021D13E4(UnkStruct_ov100_021D4DD8 *param0)
 
     {
         ov100_021D4E3C(&v0->unk_0C.unk_00, HEAP_ID_111);
-        ov100_021D4E70(&v0->unk_0C.unk_00, 0, 191, (0xffff / 192) * 2, FX32_CONST(2), 1 * 100, REG_DB_BG1HOFS_ADDR, 0, 0x1000, 1);
+        ov100_021D4E70(&v0->unk_0C.unk_00, 0, 191, (0xffff / 192) * 2, FX32_CONST(2), 1 * 100, (u32)REG_DB_BG1HOFS_ADDR, 0, 0x1000, 1);
     }
     {
         UnkStruct_ov100_021D4EBC v1 = {
